Rename nimble_compare to compare_NIPA_to_IOT
Names what the package is actually used for. Module path, test file name, the
ruff per-file-ignore in pyproject and every internal reference move together.
No behavior change; both worked examples still reproduce their published
numbers and all 59 tests pass.

The directory could not be renamed wholesale because a csv in its gitignored
output/ was held open by another process, so the tracked files moved
individually and the old output/ is left behind to be deleted by hand.
