+++ bedrock/analysis/compare_NIPA_to_IOT/matching.py
@@ -40,8 +40,11 @@ from difflib import SequenceMatcher
 
 import pandas as pd
 
-from bedrock.analysis.nimble_compare.hierarchy import split_residual, token_relation
-from bedrock.analysis.nimble_compare.series import (
+from bedrock.analysis.compare_NIPA_to_IOT.hierarchy import (
+    split_residual,
+    token_relation,
+)
+from bedrock.analysis.compare_NIPA_to_IOT.series import (
     LabeledSeries,
     normalize_code,
     normalize_name,
